Move the classes to a package
